Small fixes for travis

+++ fuzzer/main.cpp
@@ -167,7 +167,7 @@ int main() {
 
     std::clog << std::endl;
     for (size_t iteration = 0;; iteration++) {
-        std::size_t len = dist50(rng) + 3;
+        std::size_t len = static_cast<std::size_t>(dist50(rng)) + 3;
 
         for (auto clip_type : { mapbox::geometry::wagyu::clip_type_union,
                                 mapbox::geometry::wagyu::clip_type_intersection,
@@ -188,8 +188,8 @@ int main() {
                 while (num_rings > 0) {
                     mapbox::geometry::linear_ring<std::int64_t> ring;
                     for (std::size_t i = 0; i < len; ++i) {
-                        std::int64_t x = dist50(rng);
-                        std::int64_t y = dist50(rng);
+                        std::int64_t x = static_cast<std::int64_t>(dist50(rng));
+                        std::int64_t y = static_cast<std::int64_t>(dist50(rng));
 
                         ring.push_back({ x, y });
                     }
